Make a shared prompt with the solarized theme
Based on the colors chosen for the solarized theme, I implemented
my own prompt design, which I made common for zsh & bash.

Solarized is a sixteen color palette (eight monotones, eight accent
colors) designed for use with terminal and gui applications. It has
several unique properties. I designed this colorscheme with both
precise CIELAB lightness relationships and a refined set of hues
based on fixed color wheel relationships. It has been tested
extensively in real world use on color calibrated displays (as well as
uncalibrated/intentionally miscalibrated displays) and in a variety of
lighting conditions.

The Values

    SOLARIZED HEX     16/8 TERMCOL  XTERM/HEX   L*A*B      RGB         HSB
    --------- ------- ---- -------  ----------- ---------- ----------- -----------
    base03    #002b36  8/4 brblack  234 #1c1c1c 15 -12 -12   0  43  54 193 100  21
    base02    #073642  0/4 black    235 #262626 20 -12 -12   7  54  66 192  90  26
    base01    #586e75 10/7 brgreen  240 #585858 45 -07 -07  88 110 117 194  25  46
    base00    #657b83 11/7 bryellow 241 #626262 50 -07 -07 101 123 131 195  23  51
    base0     #839496 12/6 brblue   244 #808080 60 -06 -03 131 148 150 186  13  59
    base1     #93a1a1 14/4 brcyan   245 #8a8a8a 65 -05 -02 147 161 161 180   9  63
    base2     #eee8d5  7/7 white    254 #e4e4e4 92 -00  10 238 232 213  44  11  93
    base3     #fdf6e3 15/7 brwhite  230 #ffffd7 97  00  10 253 246 227  44  10  99
    yellow    #b58900  3/3 yellow   136 #af8700 60  10  65 181 137   0  45 100  71
    orange    #cb4b16  9/3 brred    166 #d75f00 50  50  55 203  75  22  18  89  80
    red       #dc322f  1/1 red      160 #d70000 50  65  45 220  50  47   1  79  86
    magenta   #d33682  5/5 magenta  125 #af005f 50  65 -05 211  54 130 331  74  83
    violet    #6c71c4 13/5 brmagenta 61 #5f5faf 50  15 -45 108 113 196 237  45  77
    blue      #268bd2  4/4 blue      33 #0087ff 55 -10 -45  38 139 210 205  82  82
    cyan      #2aa198  6/6 cyan      37 #00afaf 60 -35 -05  42 161 152 175  74  63
    green     #859900  2/2 green     64 #5f8700 60 -20  65 133 153   0  68 100  60

Shell 8-bit colors

As mentioned before, non-printing escape sequences have to be enclosed
in \[\033[ and \]. For colour escape sequences, they should also be
followed by a lowercase m.

If you try out the following prompts in an xterm and find that you
aren't seeing the colours named, check out your ~/.Xdefaults file (and
possibly its bretheren) for lines like XTerm*Foreground:
BlanchedAlmond. This can be commented out by placing an exclamation
mark ("!") in front of it. Of course, this will also be dependent on
what terminal emulator you're using. This is the likeliest place that
your term foreground colours would be overridden.

To include blue text in the prompt:

    PS1="\[\033[34m\][\$(date +%H%M)][\u@\h:\w]$ "

The problem with this prompt is that the blue colour that starts with
the 34 colour code is never switched back to the regular colour, so any
text you type after the prompt is still in the colour of the prompt.
This is also a dark shade of blue, so combining it with the bold code
might help:

    PS1="\[\033[1;34m\][\$(date +%H%M)][\u@\h:\w]$\[\033[0m\] "

The prompt is now in light blue, and it ends by switching the colour
back to nothing (whatever foreground colour you had previously).

    Black       0;30     Dark Gray     1;30
    Blue        0;34     Light Blue    1;34
    Green       0;32     Light Green   1;32
    Cyan        0;36     Light Cyan    1;36
    Red         0;31     Light Red     1;31
    Purple      0;35     Light Purple  1;35
    Brown       0;33     Yellow        1;33
    Light Gray  0;37     White         1;37

See https://tldp.org/HOWTO/Bash-Prompt-HOWTO/x329.html