Fix incomprehensible terminal output
When scrolling through the command history, incomprehensible artifacts
appeared at the beginning of the line and this caused great
inconvenience when working with the terminal, it turned out to be fixed
in this way. Perhaps this is not the best option, but it works and you
can do more useful things.