Redo the installation functions
Reworked the installation process and described the documentation for
the new commands. Among the old teams, not all are fully working and
because of this they give false hope that some problems will be fixed.
For now, I have left only the installation functions that are regularly
checked on CI and should work stably.

Since logging changes interfered with the normal operation of ci
checks, logging was completely removed.