Set push.autoSetupRemote=true for git config
This is a very convenient setup for me.

push.autoSetupRemote
If set to "true" assume --set-upstream on default push when no upstream
tracking exists for the current branch; this option takes effect with
push.default options simple, upstream, and current. It is useful if by
default you want new branches to be pushed to the default remote (like
the behavior of push.default=current) and you also want the upstream
tracking to be set. Workflows most likely to benefit from this option
are simple central workflows where all branches are expected to have
the same name on the remote.